+++ libs/qCC_db/ccSubMesh.h
@@ -47,7 +47,9 @@ public:
 	ccGenericPointCloud* getAssociatedCloud() const override;
 	void refreshBB() override;
 	bool interpolateNormals(unsigned triIndex, const CCVector3& P, CCVector3& N) override;
+	bool interpolateNormalsBC(unsigned triIndex, const CCVector3d& w, CCVector3& N) override;
 	bool interpolateColors(unsigned triIndex, const CCVector3& P, ccColor::Rgb& rgb) override;
+	bool interpolateColorsBC(unsigned triIndex, const CCVector3d& w, ccColor::Rgb& rgb) override;
 	bool getColorFromMaterial(unsigned triIndex, const CCVector3& P, ccColor::Rgb& rgb, bool interpolateColorIfNoTexture) override;
 	bool getVertexColorFromMaterial(unsigned triIndex, unsigned char vertIndex, ccColor::Rgb& rgb, bool returnColorIfNoTexture) override;
 	bool hasMaterials() const override;